Refactor metric functions: rename metrics module and update imports for consistency
